refactor: store 구조 변경
